feat(632): extract logic from function file to util
